requirements.txt
click==6.6
Django==1.10.4
django-rq==0.9.4
python-dateutil==2.6.0
python-env-utils==0.2.0
PyYAML==3.12
redis==2.10.5
rq==0.7.1
six==1.10.0
